ai/requirements.txt
Flask==3.0.0
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
prophet==1.1.5
plotly==5.24.1
mysql-connector-python==8.4.0
requests==2.32.3
